refactor: logger replaced to return error
