+++ backend/internal/migrate/migrate.go
@@ -1,13 +1,11 @@
-package app
+package migrate
 
 import (
-	"context"
 	"database/sql"
 
 	"github.com/golang-migrate/migrate/v4"
 	psql "github.com/golang-migrate/migrate/v4/database/postgres"
 	_ "github.com/golang-migrate/migrate/v4/source/file"
-	"github.com/kotopesp/sos-kotopes/pkg/logger"
 	_ "github.com/lib/pq"
 )
 
@@ -17,23 +15,25 @@ const (
 	pathToMigrations = "file:///app/internal/data/"
 )
 
-func migrateUp(ctx context.Context, dataBaseURL string) error {
+func Up(dataBaseURL string) error {
 
 	db, err := sql.Open(dataBase, dataBaseURL+sslMode)
 	if err != nil {
-		logger.Log().Fatal(ctx, err.Error())
+		return err
 	}
 
 	driver, err := psql.WithInstance(db, &psql.Config{})
 	if err != nil {
-		logger.Log().Fatal(ctx, err.Error())
+		return err
 	}
 
 	m, err := migrate.NewWithDatabaseInstance(
 		pathToMigrations,
-		dataBase, driver)
+		dataBase,
+		driver,
+	)
 	if err != nil {
-		logger.Log().Fatal(ctx, err.Error())
+		return err
 	}
 
 	if err := m.Up(); err != migrate.ErrNoChange {
